Split Up Factory
- Refactored Factory Component into seperate TSX Files to prevent needless imports / dependencies as part of Tech Debt #26
