+++ frontend/components/GenerateTemplate/MainContent/Editor/EditorBlock.tsx
@@ -7,14 +7,8 @@ import {
   PlusIcon,
   Trash2Icon,
 } from "lucide-react"
-import {
-  Control,
-  Controller,
-  FieldValues,
-  useFieldArray,
-} from "react-hook-form"
+import { Control, FieldValues } from "react-hook-form"
 
-import { AutosizeTextarea } from "@/components/ui/autosize-textarea"
 import { Button } from "@/components/ui/button"
 import {
   Card,
@@ -23,8 +17,8 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card"
-import { Checkbox } from "@/components/ui/checkbox"
-import { Input } from "@/components/ui/input"
+
+import InputGenerator from "./InputGenerator"
 
 const EditorBlock = ({
   block,
@@ -105,181 +99,4 @@ const EditorBlock = ({
   )
 }
 
-const InputGenerator: React.FC<{
-  variables: IVariable
-  // register: UseFormRegister<any>
-  control: Control<any>
-}> = ({ variables, control }) => {
-  const { fields, append, remove } = useFieldArray({
-    control,
-    name: variables.name,
-  })
-
-  switch (variables._type) {
-    case "input":
-      return <InputField variables={variables} control={control} />
-
-    case "textArea":
-      return <TextField variables={variables} control={control} />
-
-    case "checkBox":
-      return <CheckBoxField variables={variables} control={control} />
-
-    case "list":
-      return <ListField variables={variables} control={control} />
-    case "object":
-      return <ObjectField variables={variables} control={control} />
-    default:
-      return (
-        <h1 className="pb-2 font-semibold text-red-600">
-          Non-Supported Type: {variables._type}
-        </h1>
-      )
-  }
-}
-
 export default EditorBlock
-
-interface IListFieldProps {
-  variables: IVariable
-  control: Control
-}
-
-const InputField: React.FC<IListFieldProps> = ({ variables, control }) => {
-  return (
-    <Controller
-      control={control}
-      name={variables.name}
-      defaultValue={variables.defaultValue as string}
-      render={({ field }) => (
-        <div className="w-full pb-4">
-          <h4 className="pb-2 font-semibold">{variables.label}</h4>
-          <Input {...field} name={variables.name} />
-        </div>
-      )}
-    />
-  )
-}
-
-const TextField: React.FC<IListFieldProps> = ({ variables, control }) => {
-  return (
-    <Controller
-      control={control}
-      name={variables.name}
-      defaultValue={variables.defaultValue as string}
-      render={({ field }) => (
-        <div className="w-full pb-4">
-          <h4 className="pb-2 font-semibold">{variables.label}</h4>
-          <AutosizeTextarea {...field} name={variables.name} />
-        </div>
-      )}
-    />
-  )
-}
-
-const CheckBoxField: React.FC<IListFieldProps> = ({ variables, control }) => {
-  return (
-    <Controller
-      control={control}
-      name={variables.name}
-      defaultValue={variables.defaultValue as boolean}
-      render={({ field }) => (
-        <div className="flex w-full items-center justify-start pb-4">
-          <h4 className="pr-2 font-semibold">{variables.label}</h4>
-
-          <Checkbox
-            {...field}
-            name={variables.name}
-            checked={field.value}
-            onCheckedChange={(isChecked) => field.onChange(isChecked)}
-          />
-        </div>
-      )}
-    />
-  )
-}
-
-const ObjectField: React.FC<IListFieldProps> = ({ variables, control }) => {
-  return (
-    <div className="pb-4">
-      <h4 className="pb-2 font-semibold">{variables.label}</h4>
-
-      <div className="flex w-full gap-4">
-        {variables.objectSchema &&
-          variables.objectSchema.map((schema, index) => {
-            let defaultValue: any =
-              variables.defaultValue?.[
-                schema.name as keyof typeof variables.defaultValue
-              ] ?? ""
-
-            return (
-              <InputGenerator
-                variables={{
-                  name: `${variables.name}.${schema.name}`,
-                  defaultValue: defaultValue as string | boolean | string[],
-                  _type: schema._type,
-                  label: schema.label,
-                }}
-                control={control}
-                key={index}
-              />
-            )
-          })}
-      </div>
-    </div>
-  )
-}
-
-const ListField: React.FC<IListFieldProps> = ({ variables, control }) => {
-  const { fields, append, remove } = useFieldArray({
-    control,
-    name: variables.name,
-  })
-
-  return (
-    <div className="pb-4">
-      <h4 className="pb-2 font-semibold">{variables.label}</h4>
-
-      <div className="">
-        {fields.map((field, index) => (
-          <div className="flex w-full gap-4" key={field.id}>
-            {variables.listSchema &&
-              variables.listSchema.map((schema, subIndex) => {
-                let defaultValue: any =
-                  variables.defaultValue?.[
-                    index as keyof typeof variables.defaultValue
-                  ]?.[schema.name] ?? ""
-
-                return (
-                  <InputGenerator
-                    variables={{
-                      name: `${variables.name}[${index}].${schema.name}`,
-                      defaultValue: defaultValue as string | boolean | string[],
-                      _type: schema._type,
-                      label: schema.label,
-                    }}
-                    control={control}
-                    key={`${index}-${subIndex}`}
-                  />
-                )
-              })}
-            <div className="mt-8">
-              <Button variant={"outline"} onClick={() => remove(index)}>
-                Delete
-              </Button>
-            </div>
-          </div>
-        ))}
-      </div>
-
-      <Button
-        variant={"outline"}
-        onClick={() => {
-          append({})
-        }}
-      >
-        Add
-      </Button>
-    </div>
-  )
-}
