+++ frontend/components/GenerateTemplate/MainContent/Editor/types.ts
@@ -1,0 +1,7 @@
+import { IVariable } from "@/api/generated"
+import { Control } from "react-hook-form"
+
+export interface IListFieldProps {
+    variables: IVariable
+    control: Control
+}
