+++ frontend/components/GenerateTemplate/MainContent/Preview.tsx
@@ -1,4 +1,3 @@
-import { LegacyRef } from "react"
 import ReactMarkdown from "react-markdown"
 import rehypeRaw from "rehype-raw"
 import remarkGfm from "remark-gfm"
